requirements.txt
future
emoji
sqlalchemy
beautifulsoup4
requests
python-telegram-bot==11.1.0
psycopg2-binary
feedparser
geopy
nltk
psutil
aiohttp>=2.2.5
googletrans
jikanpy
